Move worker.workon to cli.hunt.workon
Why:

Now that cli hunt is using ExperimentClient.workon there is no need for
code in worker. This code is only relevant for cli.hunt.
This also solves a circular import issue.
